refactor shared controllers
