+++ lib/src/features/library/presentation/widgets/item_list_tile_widget.dart
@@ -1,5 +1,4 @@
 import 'package:book_adapter/src/constants/constants.dart';
-import 'package:book_adapter/src/controller/storage_controller.dart';
 import 'package:book_adapter/src/features/library/data/book_item.dart';
 import 'package:book_adapter/src/features/library/data/item.dart';
 import 'package:book_adapter/src/features/library/data/series_item.dart';
@@ -8,6 +7,7 @@ import 'package:book_adapter/src/features/library/model/book_status_notifier.dar
 import 'package:book_adapter/src/features/library/presentation/library_view_controller.dart';
 import 'package:book_adapter/src/features/library/presentation/widgets/default_item_image.dart';
 import 'package:book_adapter/src/features/reader/current_book.dart';
+import 'package:book_adapter/src/shared/controller/storage_controller.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
