+++ lib/src/features/profile/edit_profile_view.dart
@@ -1,9 +1,9 @@
 // Handles Profile edit action
 
 import 'package:book_adapter/src/constants/constants.dart';
-import 'package:book_adapter/src/controller/firebase_controller.dart';
 import 'package:book_adapter/src/features/profile/edit_profile_view_controller.dart';
 import 'package:book_adapter/src/features/profile/widgets/profile_widget.dart';
+import 'package:book_adapter/src/shared/controller/firebase_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
