+++ lib/src/features/profile/edit_profile_view_controller.dart
@@ -1,4 +1,4 @@
-import 'package:book_adapter/src/controller/firebase_controller.dart';
+import 'package:book_adapter/src/shared/controller/firebase_controller.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final editProfileViewController =
