+++ lib/src/model/user_model.dart
@@ -1,8 +1,8 @@
-import 'package:book_adapter/src/controller/firebase_controller.dart';
 import 'package:book_adapter/src/exceptions/app_exception.dart';
 import 'package:book_adapter/src/features/auth/data/user_data.dart';
 import 'package:book_adapter/src/features/library/data/book_item.dart';
 import 'package:book_adapter/src/model/queue_model.dart';
+import 'package:book_adapter/src/shared/controller/firebase_controller.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final userModelProvider =
